# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Go + Gin + Redis + RocketMQ 的高并发秒杀系统",
        "title": "秒杀系统 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/activity/{id}": {
            "get": {
                "description": "根据活动 ID 查询秒杀活动详情（三级缓存）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "查询秒杀活动详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "活动ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/seed": {
            "post": {
                "description": "一键创建测试商品和秒杀活动并自动预热（开发环境使用）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "创建测试数据",
                "responses": {
                    "200": {
                        "description": "测试数据创建成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/warmup": {
            "post": {
                "description": "将指定秒杀活动的库存加载到 Redis 缓存中（布隆过滤器 + 库存 Key）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理后台"
                ],
                "summary": "秒杀活动预热",
                "parameters": [
                    {
                        "description": "预热请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.WarmUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "预热成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/order/{orderNo}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据订单号查询订单详情（仅能查看自己的订单）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "查询订单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "orderNo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/orders": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "分页查询当前登录用户的订单列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "查询订单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（默认1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认10，最大50）",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/product/{id}": {
            "get": {
                "description": "根据商品 ID 查询商品详情（三级缓存：L1 本地 → L2 Redis → L3 MySQL）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品"
                ],
                "summary": "查询商品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "商品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/seckill": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "提交秒杀下单请求（异步处理），需要先登录获取 Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "秒杀"
                ],
                "summary": "秒杀下单",
                "parameters": [
                    {
                        "description": "秒杀请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.SeckillRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "抢购成功，排队处理中",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/seckill/result": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "轮询查询秒杀下单结果（SUCCESS / PROCESSING / FAILED）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "秒杀"
                ],
                "summary": "查询秒杀结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "活动ID",
                        "name": "activity_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "商品ID",
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "description": "使用用户名和密码登录，返回 JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功，返回 token 和 user_id",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/register": {
            "post": {
                "description": "通过用户名和密码注册新用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handler.RegisterRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 6
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                }
            }
        },
        "handler.SeckillRequest": {
            "type": "object",
            "required": [
                "activity_id",
                "product_id"
            ],
            "properties": {
                "activity_id": {
                    "description": "秒杀活动 ID",
                    "type": "integer"
                },
                "product_id": {
                    "description": "商品 ID",
                    "type": "integer"
                }
            }
        },
        "handler.WarmUpRequest": {
            "type": "object",
            "required": [
                "activity_id"
            ],
            "properties": {
                "activity_id": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer JWT Token（格式：Bearer eyJhbG...)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}